change to 12 bit samples

+++ inc/output.h
@@ -7,8 +7,8 @@
 
 typedef struct output_state_s {
   uint32_t period;
-  uint8_t* region_one;
-  uint8_t* region_two;
+  uint16_t* region_one;
+  uint16_t* region_two;
   size_t output_buffer_length;
 } output_state;
 
@@ -16,7 +16,7 @@ typedef struct output_state_s {
 bool output_init(output_state* o);
 
 //output_get_buffer - returns the back buffer or NULL if the back buffer is filled
-uint8_t* output_get_buffer(output_state* o);
+uint16_t* output_get_buffer(output_state* o);
 
 //output_set_filled - set the back buffer as filled with fill_length samples
 bool output_set_filled(output_state* o, size_t fill_length);
